Split out bound and multiplier
